requirements.txt
appdirs==1.4.3
argcomplete==1.9.3
attrs==19.1.0
black==19.3b0
boto3==1.9.237
botocore==1.12.237
bravado-core==5.13.2
certifi==2019.9.11
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
docutils==0.15.2
durationpy==0.5
entrypoints==0.3
flake8==3.7.8
flasgger==0.9.3
Flask==1.1.1
future==0.16.0
hjson==3.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
jsonref==0.2
jsonschema==2.6.0
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack-python==0.5.6
placebo==0.9.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyrsistent==0.15.4
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rfc3987==1.3.8
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.1
Unidecode==1.1.1
urllib3==1.25.6
webcolors==1.10
Werkzeug==0.16.0
wsgi-request-logger==0.4.6
zappa==0.48.2
